+++ spec/savon/options_spec.rb
@@ -255,6 +255,15 @@ describe "Options" do
     end
   end
 
+  context "global :ssl_version" do
+    it "sets the SSL version to use" do
+      HTTPI::Auth::SSL.any_instance.expects(:ssl_version=).with(:SSLv3)
+
+      client = new_client(:endpoint => @server.url, :ssl_version => :SSLv3)
+      client.call(:authenticate)
+    end
+  end
+
   context "global :ssl_verify_mode" do
     it "sets the verify mode to use" do
       HTTPI::Auth::SSL.any_instance.expects(:verify_mode=).with(:none)